Move createHtml function into `helpers.js`
