+++ packages/11ty/content/_assets/javascript/application/helper.js
@@ -195,3 +195,17 @@ export const toggleFullscreen = (
     });
   }
 };
+
+/**
+ * Paul Frazee's easy templating function
+ * https://twitter.com/pfrazee/status/1223249561063477250?s=20
+ */
+export const createHtml = (tag, attributes, ...children) => {
+  const element = document.createElement(tag);
+  for (let attribute in attributes) {
+    if (attribute === 'className') element.className = attributes[attribute];
+    else element.setAttribute(attribute, attributes[attribute]);
+  }
+  for (let child of children) element.append(child);
+  return element;
+}
